docs : Delete MainMenu folder and Separate folder
Delete MainMenues folder and Separate each component from folder
