+++ src/component/Department/Department.style.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import '../../../assets/fonts/font.css'
+import '../../assets/fonts/font.css'
 import { NavLink } from 'react-router-dom'
 export const StyledLink = styled(NavLink)`
     text-decoration : none;
