+++ src/component/LectureRanking/LectureRanking.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { LectureAndOrder,LectureOrder,AverageScore, Professor, LectureNameAndProfessor,Lecture,LectureWrapper,ComponentName, DepartmentName, DepartmentSelector, LectureRankingHeader, LectureRankingWrapper, RankingWrapper, LectureContainer } from './LectureRanking.style'
-import {useGetLecturesQuery} from '../../../api/hangangLecture'
-import {DEPARTMENT_LIST} from '../../../static/indexPage/departmentList'
+import {useGetLecturesQuery} from '../../api/hangangLecture'
+import {DEPARTMENT_LIST} from '../../static/indexPage/departmentList'
 function LectureRanking(){
     const [selectedDepartment,setDepartment] = useState({title:'교양',id:10})
     const {data, error, isLoading} = useGetLecturesQuery(selectedDepartment.id)
