+++ src/component/MainMenues/MainPageMeues.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import CurrentLecture from './CurrentLecture/CurrentLecture'
-import TimeTable from './TimeTable/TimeTable'
-import RecommendationLectureMaterial from './RecommendationLecutreMaterial/RecommendationLectureMaterial'
-import LectureRanking from './LectureRanking/LectureRanking'
+import CurrentLecture from '../CurrentLecture/CurrentLecture'
+import TimeTable from '../TimeTable/TimeTable'
+import RecommendationLectureMaterial from '../RecommendationLecutreMaterial/RecommendationLectureMaterial'
+import LectureRanking from '../LectureRanking/LectureRanking'
 import { MainMenuBrowserWrapper,FifthMaterial, FirstFourthThirdFifthWrapper, FirstMaterial, FourthMaterial, FourthThirdFifthWrapper, MainPageMenuWrapper, SecondMaterial, ThirdMaterial, ThridFifthWrapper } from './MainPageMenues.style'
-import DepartmentBrowser from './Department/Department'
+import DepartmentBrowser from '../Department/Department'
 function MainPageMenu(){
     return(
         <MainPageMenuWrapper>
